Update Montagetemporal.py
In this example, we first include the Wowza Player JavaScript API script tag in the head of the HTML file. We then create a div element with an ID of player where we will insert the video player.  In the script tag, we create a playerElement variable by selecting the player element using document.getElementById(). We then create a new player instance using the WowzaPlayer.create() method, passing in the playerElement and an object with the player configuration options.  In this case, we specify the Wowza Player license key and the source URL of the live stream we want to play. Note that you would need to replace the license key and source URL with your own values.  Once the player is created, it will automatically start playing the live stream. You can customize the player appearance and behavior using additional configuration options and the Wowza Player API methods.